# openapi.json
{"openapi": "3.1.0", "info": {"title": "Marketplace API", "version": "0.1.0"}, "paths": {"/api/v1/public/products": {"get": {"tags": ["Public"], "summary": "Get Products", "description": "Retrieve a paginated list of products for infinite scroll.\nComputes 'nearest_delivery_date' locally from early relation load.", "operationId": "get_products_api_v1_public_products_get", "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 10, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse_ProductListItem_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/public/products/{product_id}": {"get": {"tags": ["Public"], "summary": "Get Product Details", "description": "Get detailed product layout, incorporating explicitly sorted embedded offers.", "operationId": "get_product_details_api_v1_public_products__product_id__get", "parameters": [{"name": "product_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Product Id"}}, {"name": "offers_sort", "in": "query", "required": false, "schema": {"enum": ["price", "delivery_date"], "type": "string", "description": "Sort offers criteria", "default": "price", "title": "Offers Sort"}, "description": "Sort offers criteria"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductDetailResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/admin/auth/login": {"post": {"tags": ["AdminAuth"], "summary": "Login For Access Token", "description": "Authenticate an admin user and return a JWT token.\n\nVerifies credentials against the database using bcrypt hashing.", "operationId": "login_for_access_token_api_v1_admin_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest", "description": "Admin credentials for authentication"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "401": {"description": "Incorrect username or password"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/admin/products/": {"get": {"tags": ["AdminProducts"], "summary": "List Products", "description": "Retrieve list of products for administration.", "operationId": "list_products_api_v1_admin_products__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 50, "title": "Limit"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Offset"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse_ProductResponse_"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["AdminProducts"], "summary": "Create Product", "operationId": "create_product_api_v1_admin_products__post", "security": [{"HTTPBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/admin/products/{product_id}": {"get": {"tags": ["AdminProducts"], "summary": "Read Product", "operationId": "read_product_api_v1_admin_products__product_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "product_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Product Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductDetailResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["AdminProducts"], "summary": "Update Product", "operationId": "update_product_api_v1_admin_products__product_id__put", "security": [{"HTTPBearer": []}], "parameters": [{"name": "product_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Product Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["AdminProducts"], "summary": "Delete Product", "operationId": "delete_product_api_v1_admin_products__product_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "product_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Product Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/admin/products/{product_id}/image": {"post": {"tags": ["AdminProducts"], "summary": "Upload Product Image", "operationId": "upload_product_image_api_v1_admin_products__product_id__image_post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "product_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Product Id"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_product_image_api_v1_admin_products__product_id__image_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImageUploadResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/admin/sellers/": {"get": {"tags": ["AdminSellers"], "summary": "List Sellers", "description": "List all sellers for administrative purposes.", "operationId": "list_sellers_api_v1_admin_sellers__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/SellerResponse"}, "type": "array", "title": "Response List Sellers Api V1 Admin Sellers  Get"}}}}}, "security": [{"HTTPBearer": []}]}, "post": {"tags": ["AdminSellers"], "summary": "Create Seller", "description": "Create a new seller.", "operationId": "create_seller_api_v1_admin_sellers__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SellerCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SellerResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/admin/sellers/{seller_id}": {"get": {"tags": ["AdminSellers"], "summary": "Get Seller", "description": "Get details of a specific seller.", "operationId": "get_seller_api_v1_admin_sellers__seller_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "seller_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Seller Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SellerResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["AdminSellers"], "summary": "Update Seller", "description": "Update a specific seller.", "operationId": "update_seller_api_v1_admin_sellers__seller_id__put", "security": [{"HTTPBearer": []}], "parameters": [{"name": "seller_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Seller Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SellerUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SellerResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["AdminSellers"], "summary": "Delete Seller", "description": "Delete a specific seller.", "operationId": "delete_seller_api_v1_admin_sellers__seller_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "seller_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Seller Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/admin/offers/": {"post": {"tags": ["AdminOffers"], "summary": "Create Offer", "operationId": "create_offer_api_v1_admin_offers__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/OfferCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminOfferResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/admin/offers/{offer_id}": {"get": {"tags": ["AdminOffers"], "summary": "Read Offer", "operationId": "read_offer_api_v1_admin_offers__offer_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "offer_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Offer Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminOfferResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["AdminOffers"], "summary": "Update Offer", "operationId": "update_offer_api_v1_admin_offers__offer_id__put", "security": [{"HTTPBearer": []}], "parameters": [{"name": "offer_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Offer Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OfferUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminOfferResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["AdminOffers"], "summary": "Delete Offer", "operationId": "delete_offer_api_v1_admin_offers__offer_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "offer_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Offer Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"tags": ["health"], "summary": "Health Check", "description": "Root health check for load balancers and orchestrators.", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"AdminOfferResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "product_id": {"type": "string", "format": "uuid", "title": "Product Id"}, "seller_id": {"type": "string", "format": "uuid", "title": "Seller Id"}, "price": {"$ref": "#/components/schemas/Money"}, "delivery_date": {"type": "string", "format": "date", "title": "Delivery Date"}}, "type": "object", "required": ["id", "product_id", "seller_id", "price", "delivery_date"], "title": "AdminOfferResponse"}, "Body_upload_product_image_api_v1_admin_products__product_id__image_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_upload_product_image_api_v1_admin_products__product_id__image_post"}, "CurrencyCode": {"type": "string", "enum": ["RUB", "USD", "EUR", "KZT", "KGS"], "title": "CurrencyCode"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ImageUploadResponse": {"properties": {"image_url": {"type": "string", "title": "Image Url"}, "thumbnail_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Thumbnail Url"}, "object_key": {"type": "string", "title": "Object Key"}}, "type": "object", "required": ["image_url", "object_key"], "title": "ImageUploadResponse"}, "LoginRequest": {"properties": {"username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "password"], "title": "LoginRequest"}, "Money": {"properties": {"amount": {"type": "number", "minimum": 0.0, "title": "Amount"}, "currency": {"$ref": "#/components/schemas/CurrencyCode", "default": "KGS"}}, "type": "object", "required": ["amount"], "title": "Money"}, "OfferCreate": {"properties": {"seller_id": {"type": "string", "format": "uuid", "title": "Seller Id"}, "price": {"$ref": "#/components/schemas/Money"}, "delivery_date": {"type": "string", "format": "date", "title": "Delivery Date"}, "product_id": {"type": "string", "format": "uuid", "title": "Product Id"}}, "type": "object", "required": ["seller_id", "price", "delivery_date", "product_id"], "title": "OfferCreate"}, "OfferResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "seller": {"$ref": "#/components/schemas/SellerResponse"}, "price": {"$ref": "#/components/schemas/Money"}, "delivery_date": {"type": "string", "format": "date", "title": "Delivery Date"}}, "type": "object", "required": ["id", "seller", "price", "delivery_date"], "title": "OfferResponse"}, "OfferUpdate": {"properties": {"seller_id": {"anyOf": [{"type": "string", "format": "uuid"}, {"type": "null"}], "title": "Seller Id"}, "price": {"anyOf": [{"$ref": "#/components/schemas/Money"}, {"type": "null"}]}, "delivery_date": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Delivery Date"}}, "type": "object", "title": "OfferUpdate"}, "PaginatedResponse_ProductListItem_": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ProductListItem"}, "type": "array", "title": "Items"}, "next_cursor": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Next Cursor"}}, "type": "object", "required": ["items"], "title": "PaginatedResponse[ProductListItem]"}, "PaginatedResponse_ProductResponse_": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ProductResponse"}, "type": "array", "title": "Items"}, "next_cursor": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Next Cursor"}}, "type": "object", "required": ["items"], "title": "PaginatedResponse[ProductResponse]"}, "ProductAttributeCreate": {"properties": {"key": {"type": "string", "title": "Key"}, "value": {"type": "string", "title": "Value"}}, "type": "object", "required": ["key", "value"], "title": "ProductAttributeCreate"}, "ProductAttributeResponse": {"properties": {"key": {"type": "string", "title": "Key"}, "value": {"type": "string", "title": "Value"}, "id": {"type": "string", "format": "uuid", "title": "Id"}}, "type": "object", "required": ["key", "value", "id"], "title": "ProductAttributeResponse"}, "ProductCreate": {"properties": {"name": {"type": "string", "maxLength": 255, "title": "Name"}, "price": {"$ref": "#/components/schemas/Money"}, "stock": {"type": "integer", "minimum": 0.0, "title": "Stock", "default": 0}, "attributes": {"items": {"$ref": "#/components/schemas/ProductAttributeCreate"}, "type": "array", "title": "Attributes"}}, "type": "object", "required": ["name", "price"], "title": "ProductCreate"}, "ProductDetailResponse": {"properties": {"name": {"type": "string", "maxLength": 255, "title": "Name"}, "price": {"$ref": "#/components/schemas/Money"}, "stock": {"type": "integer", "minimum": 0.0, "title": "Stock", "default": 0}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "image_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Image Url"}, "thumbnail_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Thumbnail Url"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "attributes": {"items": {"$ref": "#/components/schemas/ProductAttributeResponse"}, "type": "array", "title": "Attributes"}, "offers": {"items": {"$ref": "#/components/schemas/OfferResponse"}, "type": "array", "title": "Offers"}}, "type": "object", "required": ["name", "price", "id", "created_at", "updated_at"], "title": "ProductDetailResponse"}, "ProductListItem": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "thumbnail_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Thumbnail Url"}, "price": {"$ref": "#/components/schemas/Money"}, "stock": {"type": "integer", "title": "Stock"}, "nearest_delivery_date": {"anyOf": [{"type": "string", "format": "date"}, {"type": "null"}], "title": "Nearest Delivery Date"}}, "type": "object", "required": ["id", "name", "thumbnail_url", "price", "stock"], "title": "ProductListItem"}, "ProductResponse": {"properties": {"name": {"type": "string", "maxLength": 255, "title": "Name"}, "price": {"$ref": "#/components/schemas/Money"}, "stock": {"type": "integer", "minimum": 0.0, "title": "Stock", "default": 0}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "image_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Image Url"}, "thumbnail_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Thumbnail Url"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "attributes": {"items": {"$ref": "#/components/schemas/ProductAttributeResponse"}, "type": "array", "title": "Attributes"}}, "type": "object", "required": ["name", "price", "id", "created_at", "updated_at"], "title": "ProductResponse"}, "ProductUpdate": {"properties": {"name": {"anyOf": [{"type": "string", "maxLength": 255}, {"type": "null"}], "title": "Name"}, "price": {"anyOf": [{"$ref": "#/components/schemas/Money"}, {"type": "null"}]}, "stock": {"anyOf": [{"type": "integer", "minimum": 0.0}, {"type": "null"}], "title": "Stock"}, "attributes": {"anyOf": [{"items": {"$ref": "#/components/schemas/ProductAttributeCreate"}, "type": "array"}, {"type": "null"}], "title": "Attributes"}}, "type": "object", "title": "ProductUpdate"}, "SellerCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "rating": {"type": "number", "maximum": 5.0, "minimum": 1.0, "title": "Rating"}}, "type": "object", "required": ["name", "rating"], "title": "SellerCreate"}, "SellerResponse": {"properties": {"name": {"type": "string", "title": "Name"}, "rating": {"type": "number", "maximum": 5.0, "minimum": 1.0, "title": "Rating"}, "id": {"type": "string", "format": "uuid", "title": "Id"}}, "type": "object", "required": ["name", "rating", "id"], "title": "SellerResponse"}, "SellerUpdate": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "rating": {"anyOf": [{"type": "number", "maximum": 5.0, "minimum": 1.0}, {"type": "null"}], "title": "Rating"}}, "type": "object", "title": "SellerUpdate"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "Token"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}